Move BudgetAdapter to its own file
This way the class can be more easily unit tested
